refactor: 뒤로가기로 모달 닫기 커스텀 훅으로 리팩토링
